Build Docker images for Arm v8 (#129)
* Build Docker images for Arm v8

I'm currently building this locally on Arm v8 and the build is working perfectly! Thought I'd open this up to see if you would be interested in hosting the images to save me having to rebuild for every release.

* Increase yarn network-timeout as workaround for downloading icons

This is apparently fixed in yarn v2 and there are no plans to fix it in v1. https://github.com/yarnpkg/yarn/issues/8242