Add Media component to handle rendering of media and embeds
This checks and renders media types and in the case of videos will try
the embed url as well as the normal url if one fails to play.

Videos and images render with the gallery created for the LargePost component

Embeds are sized in a 16/9 ratio since we can't fetch their size with
this implementation.

Communicates errors to parent component via callback to hide media if it
fails to play.